Removed Custom Generator Page for Now & Mix & Match
- Feature is not enabled / is broken, removed feature to prevent issues
- Removed Mix & Match from Sidebar, only showing specific page to prevent non-loading issues for generator for now
